pyproject.toml
[project]
name = "rl-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gymnasium[box2d]>=1.2.3",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "stable-baselines3>=2.7.1",
    "swig>=4.4.1",
]
